+++ tybotRoute/models/TiledeskChatbotUtil.js
@@ -517,7 +517,15 @@ class TiledeskChatbotUtil {
                 await chatbot.addParameter(TiledeskChatbotConst.REQ_CHATBOT_TOKEN, chatbotToken); // DEPRECATED
                 await chatbot.addParameter(TiledeskChatbotConst.REQ_CHATBOT_TOKEN_v2, "JWT " + chatbotToken);
             }
+            if (process.env.TILEDESK_API) {
+                await chatbot.addParameter(TiledeskChatbotConst.REQ_CHATBOT_TOKEN, chatbotToken); // DEPRECATED
+                await chatbot.addParameter(TiledeskChatbotConst.REQ_CHATBOT_TOKEN_v2, "JWT " + chatbotToken);
+            }
             
+            if (process.env.API_URL) {
+                await chatbot.addParameter(TiledeskChatbotConst.API_BASE_URL, process.env.API_URL);
+            }
+
             if (message.text && message.sender !== "_tdinternal") {
                 // await chatbot.addParameter(TiledeskChatbotConst.USER_INPUT, true); // set userInput
                 await chatbot.deleteParameter(TiledeskChatbotConst.USER_INPUT); // user wrote, delete userInput, replyv2 will not trigger timeout action
@@ -714,6 +722,7 @@ class TiledeskChatbotUtil {
                             // console.log("Adding from message.attributes:", key, "->", value);
                             await chatbot.addParameter(key, value);
                         }
+                        await chatbot.addParameter("payload", message.attributes.payload);
                     }
                     catch(err) {
                         console.error("Error importing message payload in request variables:", err);
@@ -969,7 +978,7 @@ class TiledeskChatbotUtil {
             }
           })
           .catch((error) => {
-            // console.error("An error occurred:", error);
+            console.error("(TiledeskChatbotUtil) Axios error: ", JSON.stringify(error));
             if (callback) {
               callback(error, null, null);
             }