+++ tybotRoute/tiledeskChatbotPlugs/DirectivesChatbotPlug.js
@@ -53,6 +53,7 @@ const { DirContactUpdate } = require('./directives/DirContactUpdate');
 const { DirClearTranscript } = require('./directives/DirClearTranscript');
 const { DirMoveToUnassigned } = require('./directives/DirMoveToUnassigned');
 const { DirAddTags } = require('./directives/DirAddTags');
+const { DirSendWhatsapp } = require('./directives/DirSendWhatsapp');
 
 class DirectivesChatbotPlug {
 
@@ -136,13 +137,19 @@ class DirectivesChatbotPlug {
     }
     const projectId = supportRequest.id_project;
     const tdcache = this.tdcache;
-    const tdclient = new TiledeskClient({
-      projectId: projectId,
-      token: token,
-      APIURL: API_URL,
-      APIKEY: "___",
-      log: this.log
-    });
+    let tdclient = null;
+    try {
+      tdclient = new TiledeskClient({
+        projectId: projectId,
+        token: token,
+        APIURL: API_URL,
+        APIKEY: "___",
+        log: this.log
+      });
+    }
+    catch(err) {
+      console.log("An error occurred while creating TiledeskClient in DirectivesChatbotPlug:", err);
+    }
 
     this.context =  {
       projectId: projectId,
@@ -207,7 +214,8 @@ class DirectivesChatbotPlug {
         attributes: {
           runtimeError: {
             message: message
-          }
+          },
+          subtype: "info"
         }
       }
     }
@@ -627,6 +635,17 @@ class DirectivesChatbotPlug {
       new DirWhatsappByAttribute(context).execute(directive, async (stop) => {
         let next_dir = await this.nextDirective(this.directives);
         this.process(next_dir);
+      });
+    }
+    else if (directive_name === Directives.SEND_WHATSAPP) {
+      new DirSendWhatsapp(context).execute(directive, async (stop) => {
+        if (stop == true) {
+          if (context.log) { console.log("Stoppin Actions on:", JSON.stringify(directive));}
+          this.theend();
+        } else {
+          let next_dir = await this.nextDirective(this.directives);
+          this.process(next_dir);
+        }
       });
     }
     else if (directive_name === Directives.QAPLA) {