+++ tybotRoute/tiledeskChatbotPlugs/WebhookChatbotPlug.js
@@ -22,6 +22,11 @@ class WebhookChatbotPlug {
     if (answer.attributes && answer.attributes.webhook && answer.attributes.webhook === true) {
       if (this.log) {console.log("EXECUTING WEBHOOK URL!", this.webhookurl);}
       if (this.log) {console.log("EXECUTING WEBHOOK ON CONTEXT:", JSON.stringify(context));}
+      if (!this.validWebhookURL(this.webhookurl)) {
+        if (this.log) {console.error("(WebhookChatbotPlug) Error. Invalid webhook URL:", this.webhookurl, "on context:", JSON.stringify(context));}
+        pipeline.nextplug();
+        return;
+      }
       this.execWebhook(answer, context, this.webhookurl, (err, message_from_webhook) => {
         if (this.log) {console.log("message_from_webhook:", message_from_webhook);}
         if (err) {
@@ -67,6 +72,13 @@ class WebhookChatbotPlug {
     if (this.log) {console.log("Start processing webhook...");}
   }
   
+  validWebhookURL(webhookurl) {
+    if (!webhookurl) {
+      return false;
+    }
+    return true;
+  }
+
   execWebhook(reply_message, context, webhookurl, callback) {
     if (this.log) {
       console.log("WEBHOOK. on context", JSON.stringify(context));
@@ -194,7 +206,7 @@ class WebhookChatbotPlug {
       }
     })
     .catch(function (error) {
-      console.error("An error occurred:", error);
+      console.error("(WebhookChatbotPlug) Axios error: ", JSON.stringify(error));
       if (callback) {
         callback(error, null, null);
       }