+++ tybotRoute/tiledeskChatbotPlugs/directives/DirAskGPT.js
@@ -102,7 +102,7 @@ class DirAskGPT {
     }
 
     if (!key) {
-      if (this.log) { console.log("DirGptTask - Retrieve public gptkey")}
+      if (this.log) { console.log("DirAskGPT - Retrieve public gptkey")}
       key = process.env.GPTKEY;
       publicKey = true;
     }
@@ -122,7 +122,7 @@ class DirAskGPT {
     if (publicKey === true) {
       let keep_going = await this.checkQuoteAvailability(server_base_url);
       if (keep_going === false) {
-        if (this.log) { console.log("DirGptTask - Quota exceeded for tokens. Skip the action")}
+        if (this.log) { console.log("DirAskGPT - Quota exceeded for tokens. Skip the action")}
         callback();
         return;
       }
@@ -298,7 +298,7 @@ class DirAskGPT {
         }
       })
       .catch((error) => {
-        // console.error("An error occurred:", JSON.stringify(error.data));
+        console.error("(DirAskGPT) Axios error: ", JSON.stringify(error));
         if (callback) {
           callback(error, null);
         }
@@ -316,7 +316,7 @@ class DirAskGPT {
         },
         method: "GET"
       }
-      if (this.log) { console.log("DirGptTask INTEGRATIONS_HTTPREQUEST ", INTEGRATIONS_HTTPREQUEST) }
+      if (this.log) { console.log("DirAskGPT INTEGRATIONS_HTTPREQUEST ", INTEGRATIONS_HTTPREQUEST) }
 
       this.#myrequest(
         INTEGRATIONS_HTTPREQUEST, async (err, integration) => {
@@ -348,12 +348,12 @@ class DirAskGPT {
         },
         method: "GET"
       }
-      if (this.log) { console.log("DirGptTask KB_HTTPREQUEST", KB_HTTPREQUEST); }
+      if (this.log) { console.log("DirAskGPT KB_HTTPREQUEST", KB_HTTPREQUEST); }
 
       this.#myrequest(
         KB_HTTPREQUEST, async (err, resbody) => {
           if (err) {
-            if (this.log) { console.error("DirGptTask Get kb settings error ", err); }
+            if (this.log) { console.error("DirAskGPT Get kb settings error ", err); }
             resolve(null);
           } else {
             if (!resbody.gptkey) {