+++ tybotRoute/tiledeskChatbotPlugs/directives/DirIntent.js
@@ -225,7 +225,7 @@ class DirIntent {
       }
     })
     .catch( (error) => {
-      console.error("An error occurred:", error);
+      console.error("(DirIntent) Axios error: ", JSON.stringify(error));
       if (callback) {
         callback(error, null, null);
       }