+++ tybotRoute/tiledeskChatbotPlugs/directives/DirSetAttributeV2.js
@@ -97,6 +97,11 @@ class DirSetAttributeV2 {
 
     async go(action, callback) {
         if (this.log) {console.log("(DirSetAttribute) action before filling:", JSON.stringify(action));}
+        // if (!action) {
+        //     if (this.log) {console.log("(SetAttributeV2) Error 'action' is missing");}
+        //     callback();
+        //     return;
+        // }
         if (action && !action.operation) {
             if (this.log) {console.log("(SetAttributeV2) Error operation is mandatory");}
             callback();
@@ -108,14 +113,16 @@ class DirSetAttributeV2 {
         if (this.log) { console.log("action.operation.operands.length", action.operation.operands.length); }
         if (this.log) { console.log("action.operation.operands[0].type", action.operation.operands[0].type); }
         
-        // FUN FACT: THIS TOOK A LOT OF EFFERT BUT IT WAS NEVER USED. YOU CAN SIMPLY CREATE A JSON ATTRIBUTE APPLYING
+        // FUN FACT: THIS TOOK A LOT OF EFFORT BUT IT WAS NEVER USED. YOU CAN SIMPLY CREATE A JSON ATTRIBUTE APPLYING
         // JSONparse FUNCTION TO AN ATTRIBUTE.
+        // DEPRECATED because type = json is not available in the UI!
         if (action.operation.operands && action.operation.operands.length === 1 && action.operation.operands[0].type === "json") {
             if (this.log) {console.log("(SetAttributeV2) setting json value...");}
             if (this.log) { console.log("(SetAttributeV2) setting json value... destination:", action.destination); }
             const json_value = JSON.parse(action.operation.operands[0].value);
             if (this.log) { console.log("(SetAttributeV2) json_value:", json_value); }
-            await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, action.destination, json_value);
+            await this.saveAttribute(action.destination, json_value);
+            // await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, action.destination, json_value);
             callback();
             return; // on json types no operations are permitted beyond assignment
         }
@@ -130,12 +137,12 @@ class DirSetAttributeV2 {
         //     callback();
         //     return;
         // }
-        if (action.operation.operators === undefined && action.operation.operands.length !== 1) {
+        if (action.operation?.operators === undefined && action.operation?.operands?.length !== 1) {
             if (this.log) {console.error("(DirSetAttribute) Invalid action: operators === undefined && operands.length !== 1")};
             callback();
             return;
         }
-        if (action.operation.operators !== undefined && action.operation.operators.length !== action.operation.operands.length - 1) {
+        if (action.operation?.operators !== undefined && action.operation?.operators?.length !== action.operation?.operands?.length - 1) {
             if (this.log) {console.error("(DirSetAttribute) Invalid action: operators.length !== operands.length - 1")};
             callback();
             return;
@@ -145,19 +152,87 @@ class DirSetAttributeV2 {
         //     await this.fillValues(action.operation.operands);
         // }
         // console.log("dirsetattribute, action.operation.operands:", action.operation.operands);
-        const expression = TiledeskExpression.JSONOperationToExpression(action.operation.operators, action.operation.operands);
-        const attributes = await TiledeskChatbot.allParametersStatic(this.context.tdcache, this.context.requestId);
-        // console.log("dirsetattribute, attributes:", attributes);
-        attributes.TiledeskMath = TiledeskMath;
-        attributes.TiledeskString = TiledeskString;
-        const result = new TiledeskExpression().evaluateJavascriptExpression(expression, attributes);
-        // console.log("filling in setattribute, result:", result);
-        // THE GOAL OF ATTRIBUTE-FILLING THE "DESTINATION" FIELD IS TO SUPPORT DYNAMIC ATTRIBUTES
-        // (ATTRS WHOSE NAME IS UNKNOWN AD DESIGN-TIME)
-        // STILL UNSUPPORTED IN UI
-        let destination = await this.fillDestination(action.destination);
-        await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, destination, result);
+        try {
+            const expression = TiledeskExpression.JSONOperationToExpression(action.operation?.operators, action.operation?.operands);
+            const attributes = await TiledeskChatbot.allParametersStatic(this.context.tdcache, this.context.requestId);
+            // console.log("dirsetattribute, attributes:", attributes);
+            if (attributes) {
+                attributes.TiledeskMath = TiledeskMath;
+                attributes.TiledeskString = TiledeskString;
+                const result = new TiledeskExpression().evaluateJavascriptExpression(expression, attributes);
+                // console.log("filling in setattribute, result:", result);
+                // THE GOAL OF ATTRIBUTE-FILLING THE "DESTINATION" FIELD IS TO SUPPORT DYNAMIC ATTRIBUTES
+                // (ATTRS WHOSE NAME IS UNKNOWN AD DESIGN-TIME)
+                // STILL UNSUPPORTED IN UI
+                let destination = await this.fillDestination(action.destination);
+                await this.saveAttribute(destination, result);
+            }
+        }
+        catch(err) {
+            console.error("SetAttributeV2 error:", err);
+        }
+        
+        // await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, destination, result);
         callback();
+    }
+
+    async saveAttribute(key, value, persist) {
+        if (this.log) {
+            console.log("SetAttributeV2 saving attribute:", key, value, persist);
+        }
+        await TiledeskChatbot.addParameterStatic(this.context.tdcache, this.context.requestId, key, value);
+        // if (persist) {
+        //     console.log("SetAttributeV2 persisting...");
+        //     await this.persistOnTiledesk(destination, result);
+        // }
+        
+        // await make persistent autenticato con "Chatbot Token" (solo il chatbot può usare questo servizio)
+        // sarebbe il top che accodasse
+        /**
+         * ACL ?
+         * Solo agents appartenenti a quella conversazione possono accedere agli attibuti. E gli Admin
+         * collection flow_attributes
+         * {
+         *  "projectId",
+         *  "flowId",
+         *  "requestId",
+         *  attributes: [
+         *      {"key": destination, "value": {}, 2, "Andrea"} => In pratica value è un "any"
+         *  ]
+         * }
+         */
+    }
+
+    async persistOnTiledesk(key, value) {
+        if (!process.env.PERSIST_API_ENDPOINT) {
+            return;
+        }
+        const HTTPREQUEST = {
+            url: process.env.PERSIST_API_ENDPOINT,
+            headers: {
+              'Content-Type': 'application/json',
+              'Authorization': this.fixToken(this.context.token)
+            },
+            json: json,
+            method: 'POST'
+        }
+        if (this.log) { console.log("SetAttribute. HTTPREQUEST: ", HTTPREQUEST); }
+        this.#myrequest(
+            HTTPREQUEST, async (err, resbody) => {
+                if (err) {
+                    if (this.log) {
+                        console.error("SetAttribute. persistOnTiledesk() error:", err);
+                    }
+                    // callback();
+                    // return;
+                } else {
+                    if (this.log) { console.log("SetAttribute. Attributes saved.", JSON.stringify(resbody)); }
+                    // callback();
+                    // return;
+                }
+            }
+        );
+        return;
     }
 
     async fillDestination(destination) {
@@ -255,6 +330,62 @@ class DirSetAttributeV2 {
         }
         catch(error) {
             console.error("Error while converting operands:", error);
+        }
+    }
+
+    #myrequest(options, callback) {
+        if (this.log) {
+          console.log("API URL:", options.url);
+          console.log("** Options:", JSON.stringify(options));
+        }
+        let axios_options = {
+          url: options.url,
+          method: options.method,
+          params: options.params,
+          headers: options.headers
+        }
+        if (options.json !== null) {
+          axios_options.data = options.json
+        }
+        if (this.log) {
+          console.log("axios_options:", JSON.stringify(axios_options));
+        }
+        if (options.url.startsWith("https:")) {
+          const httpsAgent = new https.Agent({
+            rejectUnauthorized: false,
+          });
+          axios_options.httpsAgent = httpsAgent;
+        }
+        axios(axios_options)
+        .then((res) => {
+            if (this.log) {
+                console.log("Response for url:", options.url);
+                console.log("Response headers:\n", JSON.stringify(res.headers));
+            }
+            if (res && res.status == 200 && res.data) {
+                if (callback) {
+                callback(null, res.data);
+                }
+            }
+            else {
+                if (callback) {
+                callback(new Error("Response status is not 200"), null);
+                }
+            }
+        })
+        .catch((error) => {
+            if (callback) {
+                callback(error, null);
+            }
+        });
+    }
+
+    fixToken(token) {
+        if (token.startsWith('JWT ')) {
+          return token
+        }
+        else {
+          return 'JWT ' + token
         }
     }
 }